requirements.txt
# Core dependencies (required for basic functionality)
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0  # For CLIP implementation via HuggingFace
Pillow>=9.0.0

# Development dependencies (uncomment to install for development work)
# jupyter>=1.0.0
# ipywidgets>=8.0.0
# matplotlib>=3.10.0
# notebook>=7.0.0
# open_clip_torch>=2.20.0  # Alternative CLIP implementation for testing
# clip>=1.0  # Original OpenAI CLIP implementation for comparison testing

# Testing dependencies (required for running tests)
pytest>=7.0.0 
pytest-asyncio>=0.21.0 
maturin==1.8.7  # For building and testing the Rust core

# Browser Use
browser-use==0.1.40
pyperclip==1.9.0
gradio==5.10.0
json-repair
langchain-mistralai==0.2.4
PyYAML>=6.0

# Optional dependencies for specific features
# requests>=2.25.0  # For additional web functionality 
